Move `src/data/balance/pokemon-evolutions/` to `src/data/`
