+++ src/data/pokemon-evolutions/gen5-pokemon-family-evolutions.ts
@@ -18,7 +18,7 @@ import {
   GENERIC_ITEM_EVO_LEVEL,
   HAPPINESS_EVO_LEVEL,
   KINGAMBIT_EVO_LEVEL,
-} from "#app/data/balance/pokemon-evolutions/enemy-pokemon-evolution-levels";
+} from "#app/data/pokemon-evolutions/enemy-pokemon-evolution-levels";
 import { MoveId } from "#enums/move-id";
 
 export const gen5pokemonFamilyEvolutions: PokemonEvolutions = {
