+++ src/data/pokemon-evolutions/gen8-pokemon-family-evolutions.ts
@@ -20,7 +20,7 @@ import {
   OVERQWIL_EVO_LEVEL,
   SIRFETCHD_EVO_LEVEL,
   SLOWPOKE_FAMILY_EVO_LEVEL,
-} from "#app/data/balance/pokemon-evolutions/enemy-pokemon-evolution-levels";
+} from "#app/data/pokemon-evolutions/enemy-pokemon-evolution-levels";
 import { MoveId } from "#enums/move-id";
 import { BiomeId } from "#enums/biome-id";
