+++ src/system/achievements.ts
@@ -7,7 +7,7 @@ import { globalScene } from "#app/global-scene";
 import { settings } from "#app/system/settings/settings-manager";
 import { ElementalType } from "#enums/elemental-type";
 import { AchvCategory } from "#enums/achv-category";
-import { pokemonEvolutions } from "#app/data/balance/pokemon-evolutions/init-pokemon-evolutions";
+import { pokemonEvolutions } from "#app/data/pokemon-evolutions/init-pokemon-evolutions";
 
 // Note: Refer to https://github.com/Despair-Games/poketernity/pull/775/files for removed achv code
 export class Achievement {
